clean up naming
