+++ .gitignore
@@ -105,3 +105,6 @@ ENV/
 
 # system
 .DS_Store
+
+# vscode
+.vscode
