+++ docs/ctparse.rst
@@ -6,7 +6,7 @@ Subpackages
 
 .. toctree::
 
-    ctparse.time
+   ctparse.time
 
 Submodules
 ----------
@@ -15,79 +15,95 @@ ctparse.corpus module
 ---------------------
 
 .. automodule:: ctparse.corpus
-    :members:
-    :undoc-members:
-    :show-inheritance:
+   :members:
+   :undoc-members:
+   :show-inheritance:
 
 ctparse.ctparse module
 ----------------------
 
 .. automodule:: ctparse.ctparse
-    :members:
-    :undoc-members:
-    :show-inheritance:
+   :members:
+   :undoc-members:
+   :show-inheritance:
 
 ctparse.loader module
 ---------------------
 
 .. automodule:: ctparse.loader
-    :members:
-    :undoc-members:
-    :show-inheritance:
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+ctparse.nb\_estimator module
+----------------------------
+
+.. automodule:: ctparse.nb_estimator
+   :members:
+   :undoc-members:
+   :show-inheritance:
 
 ctparse.nb\_scorer module
 -------------------------
 
 .. automodule:: ctparse.nb_scorer
-    :members:
-    :undoc-members:
-    :show-inheritance:
+   :members:
+   :undoc-members:
+   :show-inheritance:
 
 ctparse.partial\_parse module
 -----------------------------
 
 .. automodule:: ctparse.partial_parse
-    :members:
-    :undoc-members:
-    :show-inheritance:
+   :members:
+   :undoc-members:
+   :show-inheritance:
 
 ctparse.rule module
 -------------------
 
 .. automodule:: ctparse.rule
-    :members:
-    :undoc-members:
-    :show-inheritance:
+   :members:
+   :undoc-members:
+   :show-inheritance:
 
 ctparse.scorer module
 ---------------------
 
 .. automodule:: ctparse.scorer
-    :members:
-    :undoc-members:
-    :show-inheritance:
+   :members:
+   :undoc-members:
+   :show-inheritance:
 
 ctparse.timers module
 ---------------------
 
 .. automodule:: ctparse.timers
-    :members:
-    :undoc-members:
-    :show-inheritance:
+   :members:
+   :undoc-members:
+   :show-inheritance:
 
 ctparse.types module
 --------------------
 
 .. automodule:: ctparse.types
-    :members:
-    :undoc-members:
-    :show-inheritance:
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+ctparse.utils module
+--------------------
+
+.. automodule:: ctparse.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
 
 
 Module contents
 ---------------
 
 .. automodule:: ctparse
-    :members:
-    :undoc-members:
-    :show-inheritance:
+   :members:
+   :undoc-members:
+   :show-inheritance:
