+++ requirements_dev.txt
@@ -1,15 +1,15 @@
 pip==20.0.2
 bumpversion==0.5.3
-watchdog==0.9.0
-flake8==3.7.8
-tox==3.14.0
-coverage==4.5.4
-sphinx==2.2.0
+watchdog==0.10.2
+flake8==3.7.9
+tox==3.14.5
+coverage==5.0.4
+sphinx==2.4.4
 sphinx-rtd-theme==0.4.3
 
-twine==1.15.0
+twine==3.1.1
 
-pytest==5.2.2
-pytest-runner==5.1
+pytest==5.4.1
+pytest-runner==5.2
 pytest-cov==2.8.1
-mypy==0.740
+mypy==0.770
